+++ src/graph/node.rs
@@ -773,7 +773,7 @@ equal to one."
 mod tests {
     use bevy::prelude::*;
 
-    use crate::{EvalContext, InitContext, ParticleLayout, PropertyLayout};
+    use crate::{EvalContext, ModifierContext, ParticleLayout, PropertyLayout, ShaderWriter};
 
     use super::*;
 
@@ -796,7 +796,8 @@ mod tests {
 
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str = context.eval(&module, out).unwrap();
         assert_eq!(str, "(3.) + (2.)".to_string());
     }
@@ -819,7 +820,8 @@ mod tests {
         let out = outputs[0];
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str = context.eval(&module, out).unwrap();
         assert_eq!(str, "(3.) - (2.)".to_string());
     }
@@ -842,7 +844,8 @@ mod tests {
         let out = outputs[0];
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str = context.eval(&module, out).unwrap();
         assert_eq!(str, "(3.) * (2.)".to_string());
     }
@@ -865,7 +868,8 @@ mod tests {
         let out = outputs[0];
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str = context.eval(&module, out).unwrap();
         assert_eq!(str, "(3.) / (2.)".to_string());
     }
@@ -885,7 +889,8 @@ mod tests {
         let out = outputs[0];
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str = context.eval(&module, out).unwrap();
         assert_eq!(str, format!("particle.{}", Attribute::POSITION.name()));
     }
@@ -904,7 +909,8 @@ mod tests {
         assert_eq!(outputs.len(), 2);
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str0 = context.eval(&module, outputs[0]).unwrap();
         let str1 = context.eval(&module, outputs[1]).unwrap();
         assert_eq!(str0, format!("sim_params.{}", BuiltInOperator::Time.name()));
@@ -928,7 +934,8 @@ mod tests {
         assert_eq!(outputs.len(), 1);
         let property_layout = PropertyLayout::default();
         let particle_layout = ParticleLayout::default();
-        let mut context = InitContext::new(&property_layout, &particle_layout);
+        let mut context =
+            ShaderWriter::new(ModifierContext::Update, &property_layout, &particle_layout);
         let str = context.eval(&module, outputs[0]).unwrap();
         assert_eq!(str, "normalize(vec3<f32>(1.,1.,1.))".to_string());
     }