+++ src/modifier/position.rs
@@ -8,8 +8,8 @@ use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    calc_func_id, graph::ExprError, impl_mod_init_update, modifier::ShapeDimension, Attribute,
-    EvalContext, ExprHandle, InitContext, InitModifier, Module, UpdateContext, UpdateModifier,
+    calc_func_id, graph::ExprError, modifier::ShapeDimension, Attribute, BoxedModifier,
+    EvalContext, ExprHandle, Modifier, ModifierContext, Module, ShaderWriter,
 };
 
 /// A modifier to set the position of particles on or inside a circle/disc,
@@ -47,8 +47,6 @@ pub struct SetPositionCircleModifier {
     /// exactly equal to the radius (`dist == r`).
     pub dimension: ShapeDimension,
 }
-
-impl_mod_init_update!(SetPositionCircleModifier, &[Attribute::POSITION]);
 
 impl SetPositionCircleModifier {
     fn eval(
@@ -111,23 +109,22 @@ impl SetPositionCircleModifier {
 }
 
 #[typetag::serde]
-impl InitModifier for SetPositionCircleModifier {
-    fn apply_init(&self, module: &mut Module, context: &mut InitContext) -> Result<(), ExprError> {
-        let code = self.eval(module, context)?;
-        context.init_code += &code;
-        Ok(())
+impl Modifier for SetPositionCircleModifier {
+    fn context(&self) -> ModifierContext {
+        ModifierContext::Init | ModifierContext::Update
     }
-}
 
-#[typetag::serde]
-impl UpdateModifier for SetPositionCircleModifier {
-    fn apply_update(
-        &self,
-        module: &mut Module,
-        context: &mut UpdateContext,
-    ) -> Result<(), ExprError> {
+    fn attributes(&self) -> &[Attribute] {
+        &[Attribute::POSITION]
+    }
+
+    fn boxed_clone(&self) -> BoxedModifier {
+        Box::new(*self)
+    }
+
+    fn apply(&self, module: &mut Module, context: &mut ShaderWriter) -> Result<(), ExprError> {
         let code = self.eval(module, context)?;
-        context.update_code += &code;
+        context.main_code += &code;
         Ok(())
     }
 }
@@ -151,8 +148,6 @@ pub struct SetPositionSphereModifier {
     /// The shape dimension to set the position to.
     pub dimension: ShapeDimension,
 }
-
-impl_mod_init_update!(SetPositionSphereModifier, &[Attribute::POSITION]);
 
 impl SetPositionSphereModifier {
     fn eval(
@@ -217,23 +212,22 @@ impl SetPositionSphereModifier {
 }
 
 #[typetag::serde]
-impl InitModifier for SetPositionSphereModifier {
-    fn apply_init(&self, module: &mut Module, context: &mut InitContext) -> Result<(), ExprError> {
-        let code = self.eval(module, context)?;
-        context.init_code += &code;
-        Ok(())
+impl Modifier for SetPositionSphereModifier {
+    fn context(&self) -> ModifierContext {
+        ModifierContext::Init | ModifierContext::Update
     }
-}
 
-#[typetag::serde]
-impl UpdateModifier for SetPositionSphereModifier {
-    fn apply_update(
-        &self,
-        module: &mut Module,
-        context: &mut UpdateContext,
-    ) -> Result<(), ExprError> {
+    fn attributes(&self) -> &[Attribute] {
+        &[Attribute::POSITION]
+    }
+
+    fn boxed_clone(&self) -> BoxedModifier {
+        Box::new(*self)
+    }
+
+    fn apply(&self, module: &mut Module, context: &mut ShaderWriter) -> Result<(), ExprError> {
         let code = self.eval(module, context)?;
-        context.update_code += &code;
+        context.main_code += &code;
         Ok(())
     }
 }
@@ -270,8 +264,6 @@ pub struct SetPositionCone3dModifier {
     /// The shape dimension to set the position to.
     pub dimension: ShapeDimension,
 }
-
-impl_mod_init_update!(SetPositionCone3dModifier, &[Attribute::POSITION]);
 
 impl SetPositionCone3dModifier {
     fn eval(
@@ -335,23 +327,22 @@ impl SetPositionCone3dModifier {
 }
 
 #[typetag::serde]
-impl InitModifier for SetPositionCone3dModifier {
-    fn apply_init(&self, module: &mut Module, context: &mut InitContext) -> Result<(), ExprError> {
-        let code = self.eval(module, context)?;
-        context.init_code += &code;
-        Ok(())
+impl Modifier for SetPositionCone3dModifier {
+    fn context(&self) -> ModifierContext {
+        ModifierContext::Init | ModifierContext::Update
     }
-}
 
-#[typetag::serde]
-impl UpdateModifier for SetPositionCone3dModifier {
-    fn apply_update(
-        &self,
-        module: &mut Module,
-        context: &mut UpdateContext,
-    ) -> Result<(), ExprError> {
+    fn attributes(&self) -> &[Attribute] {
+        &[Attribute::POSITION]
+    }
+
+    fn boxed_clone(&self) -> BoxedModifier {
+        Box::new(*self)
+    }
+
+    fn apply(&self, module: &mut Module, context: &mut ShaderWriter) -> Result<(), ExprError> {
         let code = self.eval(module, context)?;
-        context.update_code += &code;
+        context.main_code += &code;
         Ok(())
     }
 }